# undergroundwebsite-vsc/my_first_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Underground Web Radio API",
        "description": "API resources",
        "contact": {
            "email": "pavlostzitzos@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/v1/producers": {
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Gets all producers",
                "description": "Gets all the producers and their details from db.",
                "operationId": "indexProducer",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Producers"
                ],
                "summary": "Creates a new producer",
                "description": "Adds a new producer and his details in db.",
                "operationId": "storeProducer",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/v1/producers/{id}": {
            "get": {
                "tags": [
                    "Producers"
                ],
                "summary": "Gets producer by id",
                "description": "Gets the producer and his details from db.",
                "operationId": "showProducer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of producer",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Producers"
                ],
                "summary": "Updates the details of a producer",
                "description": "Edit some or all the details of a producer.",
                "operationId": "updateProducer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Producers"
                ],
                "summary": "Deletes the details of a producer",
                "description": "Delete a producer.",
                "operationId": "deleteProducer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/v1/programs": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Gets all programs",
                "description": "Gets all the program slots and their details from db.",
                "operationId": "indexProgram",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Creates a new program slot",
                "description": "Adds a new program slot and it's details in db.",
                "operationId": "storeProgram",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/v1/programs/{id}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Gets program by id",
                "description": "Gets the program slot and it's details from db.",
                "operationId": "showProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of program",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Programs"
                ],
                "summary": "Updates the details of a program",
                "description": "Edit some or all the details of a program.",
                "operationId": "updateProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programs"
                ],
                "summary": "Deletes the details of a program slot",
                "description": "Delete a program.",
                "operationId": "deleteProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/v1/programs/clear": {
            "delete": {
                "tags": [
                    "Programs"
                ],
                "summary": "Deletes the entire program",
                "description": "Empties the program table.",
                "operationId": "clearProgram",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/v1/shows": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Gets all shows",
                "description": "Gets all shows and their details from db.",
                "operationId": "indexShow",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        }
    }
}